+++ ephypype/compute_inv_problem.py
@@ -15,7 +15,7 @@ from nipype.utils.filemanip import split_filename as split_f
 
 from .preproc import create_reject_dict
 from .source_space import create_MNI_label_files
-from .import_hdf5 import write_hdf5
+from .import_data import _write_hdf5
 
 
 def compute_noise_cov(cov_fname, raw):
@@ -303,7 +303,7 @@ def compute_inverse_solution(raw_filename, sbj_id, sbj_dir, fwd_filename,
 #            stc_file = op.abspath(basename + '_stc_' + str(i) + '.hdf5')
 #            write_hdf5(stc_file, stc[i].data, dataset_name='stc_data')
 
-        write_hdf5(stc_file, stc_data, dataset_name='stc_data')
+        _write_hdf5(stc_file, stc_data, dataset_name='stc_data')
 
     if ROIs_mean:
         label_ts, labels_file, label_names_file, label_coords_file = \
@@ -442,7 +442,7 @@ def compute_ROIs_inv_sol(raw_filename, sbj_id, sbj_dir, fwd_filename,
 
     from ephypype.preproc import create_reject_dict
     from ephypype.source_space import create_MNI_label_files
-    from ephypype.import_hdf5 import write_hdf5
+    from ephypype.import_data import _write_hdf5
 
     try:
         traits.undefined(events_id)
@@ -555,7 +555,7 @@ def compute_ROIs_inv_sol(raw_filename, sbj_id, sbj_dir, fwd_filename,
     if save_stc:
         for i in range(len(stc)):
             stc_file = op.abspath(basename + '_stc_' + str(i) + '.hdf5')
-            write_hdf5(stc_file, stc[i].data, dataset_name='stc_data')
+            _write_hdf5(stc_file, stc[i].data, dataset_name='stc_data')
 
     # these coo are in MRI space and we have to convert to MNI space
     labels_cortex = mne.read_labels_from_annot(sbj_id, parc=parc,
