+++ ephypype/tests/test_import_ctf.py
@@ -3,7 +3,8 @@ import mne
 import os
 
 from ephypype.aux_tools import _change_wd
-from ephypype.import_ctf import convert_ds_to_raw_fif  # noqa
+from ephypype.import_ctf import _convert_ds_to_raw_fif  # noqa
+
 
 data_path = mne.datasets.testing.data_path()
 ds_fname = os.path.join(data_path, 'CTF', 'testdata_ctf.ds')
@@ -14,5 +15,5 @@ def test_convert_ds_to_raw_fif():
 
     _change_wd()
 
-    raw_fif_file = convert_ds_to_raw_fif(ds_fname)
+    raw_fif_file = _convert_ds_to_raw_fif(ds_fname)
     mne.io.read_raw_fif(raw_fif_file)
