+++ ephypype/tests/test_import_hdf5.py
@@ -3,7 +3,7 @@
 import mne
 import os.path as op
 
-from ephypype.import_hdf5 import write_hdf5, read_hdf5
+from .import_data import _write_hdf5, _read_hdf5
 from numpy.testing import assert_array_almost_equal
 
 data_path = mne.datasets.sample.data_path()
@@ -21,9 +21,9 @@ def test_read_write_hdf5():
     data = raw.get_data(picks=picks, start=0, stop=600)
 
     hdf5_file = raw_fname.replace('.fif', '.hdf5')
-    write_hdf5(hdf5_file, data, dataset_name='data')
+    _write_hdf5(hdf5_file, data, dataset_name='data')
 
-    data_test = read_hdf5(hdf5_file, dataset_name='data')
+    data_test = _read_hdf5(hdf5_file, dataset_name='data')
 
     assert data.shape == data_test.shape
     assert_array_almost_equal(data, data_test)
